feat: add ships attribute and related methods
Added a vector, ships, that keeps track of all non-sunk ships. addShip() does as the name says and adds a ship to the ships vector. removeShip() finds the target ship and removes it from the vector, signifying it as sunk. This will be needed for placements and shooting at ships.